Changed what gives you points

+++ LD40/Assets/Scripts/Score.cs
@@ -1,4 +1,4 @@
 ﻿public class Score {
-        public static int rawScore = 0;
+        public static float rawScore = 0;
         public static int score = 0;
 }
